refactor: use namespaces when importing from the shared/ui package
