--- src/shared/ui/index.ts
@@ -1,6 +1,0 @@
-export * from "./data-display";
-export * from "./inputs";
-export * from "./layouts";
-export * from "./navigation";
-export * from "./surfaces";
-export * from "./utils";
